requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.2.0
dataclasses==0.6
docker-pycreds==0.4.0
filelock==3.3.1
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
huggingface-hub==0.0.19
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
JPype1==1.3.0
lxml==4.6.4
Markdown==3.3.6
multidict==5.2.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.4
pathtools==0.1.2
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.0
python-dateutil==2.8.2
pytorch-lightning==1.5.4
pytz==2021.3
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
sentry-sdk==1.5.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==0.6.1
tqdm==4.62.3
transformers==4.11.3
tweepy==4.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wandb==0.12.7
Werkzeug==2.0.2
xlrd==2.0.1
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
